Refactor: extract AbstractTitleMapper
