+++ recommendator/src/main/java/com/lampirg/recommendator/anidb/mal/titlemapper/SingleThreadTitleMapper.java
@@ -15,54 +15,21 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
 
 @Component
 @Qualifier("single")
 @Scope("prototype")
-public class SingleThreadTitleMapper implements TitleMapper {
-
-    MalQueryMaker queryMaker;
-    HttpEntity<String> request;
-
-    private final Map<AnimeTitle, Integer> recommendedAnime = new HashMap<>();
-    private Set<AnimeTitle> toExclude = new HashSet<>();
-
-    @Autowired
-    public void setQueryMaker(MalQueryMaker queryMaker) {
-        this.queryMaker = queryMaker;
-    }
-
-    @Override
-    public TitleMapper setRequest(HttpEntity<String> request) {
-        this.request = request;
-        return this;
-    }
-
+public class SingleThreadTitleMapper extends AbstractTitleMapper implements TitleMapper {
     @Override
     public Map<AnimeTitle, Integer> getRecommendedAnimeMap(Set<UserAnimeTitle> animeTitles) {
+        if (recommendedAnime == null)
+            recommendedAnime = new HashMap<>();
         if (!recommendedAnime.isEmpty())
             return recommendedAnime;
         for (UserAnimeTitle title : animeTitles) {
             findAndAddTitleRecommendations(title);
         }
         return recommendedAnime;
-    }
-
-    public TitleMapper fillToExclude(Set<UserAnimeTitle> toExclude) {
-        for (UserAnimeTitle title : toExclude) {
-            this.toExclude.add(title.animeTitle());
-        }
-        return this;
-    }
-
-    private void findAndAddTitleRecommendations(UserAnimeTitle title) {
-        String url = "https://api.myanimelist.net/v2/anime/"+title.animeTitle().id()+"?fields=recommendations";
-        ResponseEntity<GetAnimeDetail> response = queryMaker.exchange(url, HttpMethod.GET, request, GetAnimeDetail.class);
-        for (Recommendation recommendation : Objects.requireNonNull(response.getBody()).recommendations()) {
-            AnimeTitle animeTitle = AnimeTitle.retrieveFromMalNode(recommendation.node());
-            if (toExclude.contains(animeTitle))
-                continue;
-            recommendedAnime.merge(animeTitle, title.score(), Integer::sum);
-        }
     }
 }
